Fix achievements not saving
It was because an empty instance with a prototype inherited from the settings class would assign the values to the property of the prototype instead of the instance, as the instance lacked said property.
Now we also loop over properties that belong to the prototype and assign those values to the instance, especially as this could be an empty instance if we are loading a non-existing setting in the first place.