Split data object. start on ID references. (WIP)
Split the main data object classes into separate smaller files and will make it easier to add separate modules with their own data objects.
Made the map static so all different data object classes are now in the same collections, had to add import class for the editor due to that.
Started on the storage of data object references, this is WIP though as the whole refactoring business got out of hand, will continue on this later.