Make Discord integration way more reliable.
We now listen for errors and retry messages that returned HTTP 429, which means we send too quickly. This will also add one second to the reset time supplied by discord, because clearly that is not true at that point. My theory is that Cloudflare has their own anti-abuse burst limit set to 40 messages in succession, so even if we adhere to the rate limits, they still nuke us, thus we needed this error handling. It's not perfect as it can reorder items, but this doesn't happen very often, will revisit if it turns out to be come a problem.