Move text references into the database, referencing the enum we use to reference the actual functions.
This will show the keys are just numbers, as that is what the enum contains, so the import view is a bit... diffuse. It will have to do for now, in the future we might associate keys with a collection or enum and map those up in the editor, at least.