Add the capability to add or remove a parent from an item.
To add a parent, edit the parent and add the global item and edit it, then you get a choice to add the parent you came from as the parent for that item.
The top LI will now be marked as modified if anything has changed, including key and parent ID which is outside the JSON structure.