WIP: A second conversion of what data we populate with for referenced items
Now it's items straight from the database, we have all the data.
The key reference flag has been retired, as the key is in the items.
All data retrieval functions have been rewritten, some extra added.
All data objects have been fixed, but none of the classes, that is the WIP.