Fix editor
Solved the data format change by simply pushing the raw JSON to the editor, this has the problem of it not being remapped to data changes thoughs, so this will have to change, but for now it works so this is a commit.
The fix was done in the EditorHandler, so the incoming data is now a full DataBaseItem instead of just an instance, this means we have BOTH the JSON and reference-filled instance.