WIP Split Action code into Actions and ActionRunners
The idea is that Runners will be talking to other code, and when instantiating the classes a flag will decide if it is the Runner type or not.
This should help with splitting the data objects out so they can be used in the browser components without trying to import all future Node modules as well.