Add two new tsconfig files that will separately build client and server dist folders
Added NPM scripts to do the full build like this.
Client will build Server too due to dynamic imports, but the Server folder then gets deleted.
This is still not running, just building, the Client pages do not yet use the files in the dist_client folder. This will start happening when we have EJS going.